backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyYAML==6.0.1
openai==0.28.1
weasyprint==60.2
jinja2==3.1.2
pydantic==2.5.2
requests==2.31.0
slowapi==0.1.8
httpx==0.27.2 